+++ src/modules/bhatti/v1/nikasi-gate-pass/nikasi-gate-pass.model.ts
@@ -182,7 +182,6 @@ const NikasiGatePassSchema = new Schema<INikasiGatePass>(
     idempotencyKey: {
       type: String,
       trim: true,
-      sparse: true,
     },
   },
   {