+++ src/modules/bhatti/v1/outgoing-gate-pass/outgoing-gate-pass.model.ts
@@ -225,7 +225,6 @@ const OutgoingGatePassSchema = new Schema<IOutgoingGatePass>(
     idempotencyKey: {
       type: String,
       trim: true,
-      sparse: true,
     },
   },
   {